Start of a Plex library and headless, lightweight audio player.
WARNING: this is very early stage code and will almost certainly
completely change before it's useful.

There are two main parts to this:

* plex library - core data structures and protocol support
* player - a headless audio player

The library contains server and client parts. The server code is just
about discovery for now. More interesting is the client part ...

The client support consists of a Client type that will handle the
complexity of finding and joining a local Plex network. It's all about
the protocols involved. To actually do anything it needs to have a
Player added.

The Player is a simple interface that an application implements to
provide actual output. There is the start of a headless audio player in
the player/player.go command.

It's currently at the point where the Client will announce itself to the
network and wait for a server to attach to it. The Client will keep the
server updated about its state. The Client also forwards playback
commands from the server to the Client's Player although player.go does
very little with those commands for now.

Lots to do still!