Abstract the controller and separate controller state tracking.
This is the 1st part in an experiment to unify controllers that
subscribe and controllers that poll.

A simple `controller` interface is used to interact with the remote
controller. For now a controller is always a subscribing controller but
a polling controller will be added.

The controller's state (last seen command ID, removal timeout, etc) are
now stored alongside the controller as a "registered controller" type.
The idea is that the `Client` can do all the bookkeeping and the
controllers can be relatively simple.

We'll see how it goes.