Handle playMedia command from controller.
The Client needs to know what type of player to pass the playMedia
command to, but it seems that controllers are not required to send the
player type in the playMedia request.

So, the Client now parses the playMedia request to load the media
container from the relevant server. The player type can be inferred from
the items in the media container, after which the command can be
dispatched to the appropriate player.

This is actually a good thing since, ideally, the player should have as
little to do as possible.

One unfortunate side effect of this is the player's command channel now
contains interface{} vars, losing ay type checking. But it's good enough
for now.